refactor(ivy): do not deep import from ngtsc into ngcc (#24897)
PR Close #24897
